*: create a ServiceMonitor for CMO
This commit creates and deploys a ServiceMonitor for the Cluster
Monitoring Operator itself. The problem of monitoring the CMO if the
operator fails to bootstrap the monitoring stack is out of scope here.

This commit also adds an alerting rule for the Cluster Monitoring
Operator reconcilliation error rate as well as an alerting rule for when
the CMO is down.