Makefile: improve dependencies
This commit fixes the dependency order in the Makefile so that a single
change in, for example, a jsonnet file will result in a newly compiled
binary. These improvements also ensure that Make properly tracks
staleness to avoid re-running tasks unnecessarily.

This commit adds a new file to the assets directory that was actually a
dependency of pkg/manifests/manifests.go.