manigests,hack: only merge ClusterRoles
This commit adds extra logic to hack/merge_cluster_roles.py so that it
is able to merge lists of roles. If the script detects a resouce that is
not a ClusterRole, e.g. a Role, then it will log a warning. If the
resource is neither a Role, ClusterRole, or list of these, then it is
skipped.