Makefile: prevent which logs in stderr
This commit prevents Make from logging errors to stderr when it
is checking if it should install go-jsonnet or not. These errors are of
the form:

```
which: no jsonnet in
(/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/go/bin)
```

and are expected if jsonnet is not installed but can make it seem like
the build is failing.