Makefile: don't install binaries as direct deps
This commit modifies the Makefile dependency tree so that all utility
binaries are no longer automatically installed as direct dependencies of
the targets. This ensures that in a Docker container with no jsonnet,
for example, the entire static asset tree will not be regenerated. This
in turn allows us to avoid having to install extra tools like jq during
the container build process.