Add cluster-monitoring-view ClusterRole
This ClusterRole can be used by users to bind against, to allow viewing
the cluster monitoring stack. This is useful for usability but also to
allow us to modify the specific roles at a later point in time, because
existing users that bind to it will continue to have the correct roles.