jsonnet,assets: bump and regenerate
This commit bumps the jsonnet dependencies in order to bring in updates
to the Telemeter client. Notably, the Telemeter client now has a default
URL configured for the Telemeter server. This is configured by default
to the production URL. This can always be overridden using the Cluster
Monitoring Operator ConfigMap.