assets: bump telemeter, node-exporter
This commit bumps Telemeter client and adjusts the Jsonnet configuration
so that it matches the new layout of the project. This finally bumps
the Telemeter client ServiceMonitor to use the new serving certs CA
bundle.

This commit also pulls in the new changes for the node-exporter which
ensure that the Pod only listens on port 9100.