assets,jsonnet: fix grafana port
This commit fixes an old issue that only began to manifest itself today.
In this project we add a proxy in front of Grafana, which occupies
Grarana's normal port of 3000, and move Grarana to 3001. However, we
did not redefine the named container port `http` from 3000 -> 3001.
The result was that once upstream jsonnet for Grafana added an HTTP
health check for port 3000, on our deployments this was actually hitting
the proxy, which uses TLS, so the health check would fail.