Output correct error message with appropriate level
When cluster-monitoring-operator failed to parse configmap, it does
not output the reason why it failed. In addition, it outptus info
level.

e.g - when failed to parse configmap:
```
I0110 06:16:18.665461       1 operator.go:270] Cluster Monitoring config could not be parsed. Using defaults.
```

Due to this, trouble shooting is difficult.
This patch adds error output why it failed with appropriate level.