operator: rotate api authentication certs
Currently, when the extension api authentication certificates are
being rotated, prometheus adapter is not being updated.

This fixes it by watching the extension api authentication configmap
in the kube-system namespace in the cluster monitoring operator
directly. When this configmap is being updated (this happens, when the
CAs rotate), a new secret in the openshift-monitoring namespace is
generated with a synthesized name based on the CA hashes. Finally the
prometheus adapter deployment is being updated and rolled out.

Fixes MON-219