jsonnet,pkg: watch kubelet CA
This commit addds a new configmap-reloader container to the Prometheus
K8s pods that watch the configmap holding the CAs used to sign the
Kubelet CSRs. These CSRs are signed by the CSR controller. The CA is
rolled frequently and so must be watched by the
cluster-monitoring-operator and mirrored into the openshift-monitoring
namespace whenever it changes.