pkg/manifests: move telemeter config to env
This commit moves certain Telemeter configuration values to environment
variables rather than fields in a secret. This way, the Telemeter pod
will automatically be rolled when the configuration in the CMO configmap
is changed. Without this, the CMO would modify the Telemter secret, but
the pod was never updated because the secret fields were read using the
downward API only once during pod initialization.