pkg/tasks/prometheusadapter: synthesize secret including serving certs
Currently we only synthesize the API authentication certificates. We
should also consider the rotation of the serving certs.

This fixes it by watching the prometheus-adapter-tls secret, which is
updated by the serving certs controller.

This secret is read and also used to synthesize a consolidated TLS
secret mounted into prometheus adapter. This secret now contains all
relevant CA data for prometheus adapter.