pkg/operator: fix progressing state of cluster operator
Currently, we always set the progressing state of the cluster operator
status.

The desired behavior is to set the cluster operator status progressing
status only if the operator is not available or if the available status
has not been set at all (unknown state).

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1669258