jsonnet: programmatically remove limits
This commit adds a new utility function to remove limits from all
Deployments, DaemonSets, etc. It is a function because we can't use
`super` when computing field names for the root object in a mixin.

If we can come up with a mixin that does this, then let's do a follow
up.