pkg/client/condition: respect current and target version
Currently, we ignore the versions of the operator
and only set to progressing if the operator has been failing before.

This fixes it by also also respecting the current version of the operator.
If the target version is the same, the existing rules apply.
If the current version is different, we always set to progressing.

Fixes a part of https://bugzilla.redhat.com/show_bug.cgi?id=1690162
where the operator was launched with a new version but did not report
a progressing state.