pkg/manifests/operator: synthesize etcd secret
Currently, we do assume that a kube-etcd-client-certs secret exists
in the same namespace as the cluster monitoring stack.

The reality is etcd client certificates are created in the openshift-config
namespace, separately in a configmap for the etcd client CA and a secret including
the TLS client certificate and key.

This synthesizes the kube-etcd-client-certs secret in the monitoring namespace
or updates it when changes in the underlying etcd client CA happen.
We rely on Prometheus' ability to detect and react to those changes in-process.