alerts: Remove node interface down alert
Talked to the SDN team and this alert is too aggressive. The true
solution is that in the future the SDN team will provision an alert with
the truly interesting node networking interfaces, which also vary across
cloud environments and various other factors. For now this alert is
better left out than be a constant source of false-positives.