Ignore gosec low priority warnings
Currently there are two warnings given by gosec. These can be ignored
for the following reasons:

* Insecure verify for TLS

This comes from a component that is used for testing, hence this is not
really an issue.

* Usage of sha-1

SHA-1 is considered weak and its usage triggers a warning. However,
given that it's approved by FIPS (so it's usable by our customers), and
to not change the existing functionality, we just ignore this warning
for now. In a further iteration we can consider upgrading this to
SHA-256 or something else.