Added "code:registry_api_request_count:rate:sum" recording rule
OpenShift SRE needs to be able to track the rates of API requests to the
image registry service for Service Level Agreement compliance. The
existing "code:apiserver_request_count:rate:sum" rule doesn't
differentiate between the service each API call is going to, so it
isn't specific enough for this purpose.

To remedy this, I added another recording rule that selects only
the API call counters that correspond to the registry service, and
then performs a rate-and-sum similar to "code:apiserver_request_count:rate:sum"