alerts: etcd should still alert when a member disappears from the endpoints
There is no alert that reports when an etcd quorum member is degraded.
Add an alert that reports if any etcd member is down or potentially
unreachable. Since it is possible if the etcd metrics are collected
under a Kubernetes service and the etcd instance is removed from the
endpoint (due to implementation details of how etcd is run), extend
the alert to also capture failure rate of requests to that cluster.