jsonnet/rules: Correctly account for infrastructure usage
This changes the recording rules for CPU and memory use related to
infrastructure to fix two issues:

1. double counting of containers and pod memory/CPU accounting
2. not accounting for node level services and kernel CPU

The new algorithm is to calculate infrastructure use by subtracting
user workload (!~"openshift-.+") from cluster usage. This accounts
for everything that is not end user workloads and more accurately
reflects what belongs to cluster vs user.