manifests: Change order of OperatorGroup creation
It seems like OLM CRDs are created at the same time as the CMO, so this
might cause a race condition. As we do not need the OperatorGroup
resource until a user would potentially deploy their own prometheus-operator
via OLM, we can create the resource at the very end. This takes the
inspiration from CVO on when it creates monitoring resources.