jsonnet: node_exporter should run any available initial text exporters
There are a few collectors we wish to be present at initialization
time, such as the virt textfile collector in the openshift image.
Add a node exporter init container that executes all the textfiles
and share an emptyDir with the node exporter (which defaults to
showing textfile collection in the working dir).