jsonnet: Pod recording rules are incorrectly summing resources
In kubernetes, the sum of resources used by a pod is not the sum
of containers, but is instead the value reported for the pod
cgroup `{container=""}`. The recording rules here were incorrectly
omitting CPU usage accounted to the pod but not accounted to a
container which would under report containers that run in the pod
context like builds (but which aren't kube containers) as well as
other metrics also accounted to the pod but not containers.