test/e2e: improve e2e test resiliency
We have a couple of potential e2e flake sources:

1. We don't check the response code from the prometheus query response.

That causes the vanilla http response in case of error to be parsed as JSON which fails.
We should bail out with the http status code instead.

2. It happens that we sometimes don't pick the correct token secret.

With every service account there are two token secrets.
One is the internal Kubernetes service account token,
the other one is associated with the internal OpenShift Docker registry.
We have to pick the right one otherwise non-deterministically the wrong token will be chosen.

This fixes the above two issues.