jsonnet: Gather ~10 metrics that tell us workloads that are being used
These are the first metrics that show rough details of what is going
on for a cluster so that we can assess rough usage. It captures both
scale (number of particular workload types) and gives us a small hint
about usage type (are people using DCs or deployments, statefulsets or
jobs). The containers usage can be used to assess approx container per
pod numbers when contrasted to pod numbers, which is also useful.

For cardinality 10, we get a much better insight into whether this
cluster is in use.  We can approximate infrastructure resource count
against the user's workload resource count by looking at empty clusters,
but sheer scale helps us determine better values.