pkg/client/client.go: Fix updating deployments
Currently whenever do an update on the fetched deployment we get an
error log in the kube controller manager pod, that there was an error
syncing deployment. "please apply your changes to the latest version
 and try again". This log line happens for each deployment we update,
every resync interval.

This makes sure we have the correct deployment update in place by using
the cluster-version-operator libs as an import to ensure deployment.