Send apiserver_current_inflight_requests in telemeter.
cluster:apiserver_current_inflight_requests:sum:max_over_time:2m gives
maximum number of requests in flight over a 2-minute window. This
metric is a constant 4 time series that monitors concurrency of
kube-apiserver and openshift-apiserver with request type which can be
either 'mutating' or 'readonly'. We want to have an idea of how loaded
our api server(s) are globally.