pkg/operator/operator.go: Bump resync interval to 15 minutes
We already react on changes on any resources we need to reconcile on,
the resync interval should be there to just reconcile. Every 5 minutes
is too much, as we most of the time don't do anything.

We have been seeing a lot of CPU usage on every resync, this at least
lowers how often we spike.