add API performance dashboard
OpenShift ships with a set of read-only grafana dashboards, we
want to add an "API Performance" grafana dashboard that visualizes
kube-apiserver metrics to this set.

Benefits:
It will come to aid in troubleshooting customer cases. Customer will
be able to share with us screen shots of api error rates, request in
flight, latency, request dropped or terminated and such.