*: Bump etcd-mixin to 2b79442d8e9f
Generated with:

  $ make jsonnet/vendor  # to install jb
  $ jb --version
  v0.3.1
  $ (cd jsonnet && jb install https://github.com/coreos/etcd/Documentation/etcd-mixin)
  GET https://github.com/coreos/etcd/archive/2b79442d8e9fc54b1ac27e7e230ac0e4c132a054.tar.gz 200
  $ touch jsonnet/jsonnetfile.lock.json  # recover from any previous flubbed generation which might have left timestamps in place convincing Make it didn't need to rebuild bindata.go
  $ make generate-in-docker

This pulls in:

  $ git --no-pager log --oneline f1eca4e1fa5de962ff8079af836bb390e88d1f4c..2b79442d8e9fc54b1ac27e7e230ac0e4c132a054 -- Documentation/etcd-mixin
  2c4877064 Documentation/etcd-mixin: Use etcd_mvcc_db_total_size_in_bytes metric
  68c5f6066 Documentation/etcd-mixin: Set unique UID for Grafana dashboard
  322c38e16 Documentation/etcd-mixin: Fix etcdHighNumberOfLeaderChanges (#11448)

The first two of those are [1].  The last is [2], and as the
discussion there points out, the rate>3 approach is effectively "never
fire" (because we are unlikely to ever have more than three elections
per second).

Also interesting, is that f1eca4e1fa is in etcd's release-3.4 branch,
while I'm pinning the master branch.  The current coreos release-3.4
tip is etcd-io/etcd@31e49a4df30, which has no etcd-mixin changes since
f1eca4e1fa.  My impression is that mixin changes are unlikely to be
backported to release branches, and also unlikely to depend on the
intricacies of the underlying etcd version, so I'm tracking master
instead of release-3.4 in this commit.  The move to f1eca4e1fa had
landed here via 5c251d9893 (jsonnet/jsonnetfile.lock.json: jb update,
2020-04-17, #760).

[1]: https://github.com/etcd-io/etcd/pull/11768
[2]: https://github.com/etcd-io/etcd/pull/11448