Bug 1852767: set user workload prometheus retention
Allows the user workload prometheus retention setting to override
the cluster monitoring settings only if the cluster monitoring
setting is not set to the default.  Before this change the user workload
setting is never used because the cluster monitoring setting is
compared to an empty string instead of the default value.