test/e2e/user_workload_monitoring: have a constant amount of expected secrets.
The calculation of expected grpc secret count is potentially racy.
The current test code waits until the central grpc-tls secret is created.
However at this point of time there is no guarantee
that all derived hashed grpc secrets are also already created
as their creation is orchestrated in separate task executions which happen concurrently.

Instead, this changes it to have a const number of five (5) expected secrets,
because we know them in forehand.