Bug 1860889: decrease CMO log verbosity from 3 to 2
With `-v=3`, the logs are spammed by client-go telling that client
requests have been throttled. Looking at the Kubernetes code, we won't
lose anything useful by lowering the verbosity level.

The change also increases the QPS and burst settings (respectively 5->20
and 10->40) because the default values aren't suited for CMO that runs
lots of tasks in parallel. With QPS=100 and Burst=200, the throttling
logs disappeared but we felt that these values were too high for the API
server.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>