jsonnet: Collect all resource counts for telemetry
Transcribe all etcd_object_counts (number of all objects) into
cluster resource usage for reporting in cluster telemetry.

The data about how many instances of a CRD are in use on the cluster
is a key question for analyzing the success and health of a cluster
as well as for understanding how a cluster can be more successful.
From a configuration perspective, it indicates use of core platform
features. From an operator perspective, it shows adoption and usage
in a way that installation stats alone do not capture. From a support
perspective it highlights interactions between operators that may
help problem resolution.

On an average cluster this increases the series gathered by ~130
series (18 gathered today, 155 after this change, + 280 on average).
While this is a 50% increase in gathered series, it drives significant
value in all of the domains mentioned above and we believe we have
the budget to perform this increase and that using the budget on these
series specifically will deliver significant value.