*: fix prometheus-operator's cluster role
This is a follow-up of #970. Because of an old reference in
jsonnet/jsonnetfile.json to github.com/core/prometheus-operator, the
AlertmanagerConfig changes didn't fully trickle down to the generated
assets. The end result is that the prometheus-operator service account
didn't have permissions to manager AlertmanagerConfig objects which
eventually failed the reconciliation of Alertmanager.

In addition it also ensures that the new AlertmanagerConfig CRD is
declared in the `OperatorGroup` configuration and included in the objects
collected by `must-gather`.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>