cmd/operator: increase rest config burst and QPS
Increase rest config QPS value to reflect what is set in upstream
prometheus-operator.
Also, it was reported that CMO is producing a lot of throttling error
logs when it is run in an environnement with a lot CRDs. To fix that, we
can align with the value set in other component that encountered the
same issue. For instance, in `oc` the burst value was set to 200.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>