jsonnet/rules: Add recording rules for ingress traffic and error rate
The overall rates of ingress traffic to the cluster are relevant in
assessing health and usage - these rules add calculations for the
rate of requests that error (5xx and other types of errors) across
the frontends, the total requests passing across the frontends, the
bandwidth used in both directions, and the number of open connections.
Error rate is split for openshift and workloads (like CPU is) in order
to better categorize trends before, during, and after upgrades.

This data is relevant for both admins and engineering teams to classify
and categorize the overall health and traffic of the cluster. This adds
seven new series to telemetry via the cluster:usage: prefix.