jsonnet/rules: Add a recording rule measuring pod readiness
For all pending, running, or unknown pods report a 0 if the pod is
unready and a 1 if a pod ready. Use that to build two aggregate
average readiness fractions for openshift-* and !openshift-*
namespace pods, which we will use to measure impact of upgrades
on pod availability before, during, and after the change.

The fundamental rule can be used in the future for alerting when
the aggregate readiness drops below a threshold after we identify
a minimum bar.

This adds 2 series to telemetry for monitoring the impact of
upgrades.