jsonnet: Ingress recording rule for error fraction is incorrect
The denominator of the recording rule should be `increase`, like
the numerator, instead of `irate`. `increase` is more stable over
the intervals we are testing. Introduced due to a copy and paste
error as the rule was being refactored, and when assessing the
impact in 4.7 clusters I realized we had fractions > 1 showing up.