jsonnet/rules: Mask KubeDeploymentReplicasMismatch alert for upgrade
The KubeDeploymentReplicasMismatch alert fires after 15m, but
deployments that run on all three control plane nodes may see
disruption that lasts for 3 x 20m = 1h during upgrades of slow to
reboot metal hardware. Instead, prevent the alert rule from firing
when a control plane nodes are unready (which should be covered
by other alerts). While this may mask failures that occur on non
control plane nodes, once the control plane is recovered that
alert will begin to fire. Provide the alert with a more accurate
description.

A future alert may be able to leverage a better metric such as
"1 if and only if the control plane is not being upgraded and has
been progressing for less than X budget per machine", but that would
only be slightly less accurate and simpler may be better.