jsonnet/rules: TargetDown should exclude unschedulable nodes
The TargetDown alert fires today when multiple nodes are upgraded
in sequence and one or more daemonsets are successively out for
longer than 10m. Since upgrade is normal behavior and we know that
nodes will be upgrading continuously as the control plane rolls
out, alter the alert to exclude unschedulable nodes (a reasonably
accurate proxy for a node being upgraded today that can be improved
later), wait a longer period of time before firing (65m of
continuous availability), and describe the alert in more detail.

TargetDown is still useful as a bedrock symptom-based alert for
target disruption since it is likely to catch a very wide range
of problems related to running in Kubernetes, so we don't restrict
it further. In investigation this alert fires commonly in clusters
(90% of fires last longer than 60m or even 1d) so increasing the
required trigger time will still provide that backstop, while
removing higher frequency noise from normal cluster operations.