jsonnet: make AggregatedAPIDown more resilient
According to what we've seen in CI, it might take more time for the
apiservice to claim the aggregated API during upgrades than what has
been set in the upstream alert. It might take up to 15 minutes for the
aggregated API to be claimed during upgrades. Thus, we need to increase
the `for` clause to 15m to reflect that.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>