generate: use jsonnet v0.16.0
Previously, the version of the go-jsonnet package that we were
installing in the docker image used to generate manifests wasn't fixed
to a particular version. This resulted in golang errors with go-jsonnet
v0.17.0 as it required golang 1.10, but we are using a 1.9 image.

Also, this required to get the package from the jsonnet repository
instead of go-jsonnet as the assets weren't available for go-jsonnet.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>