hack: add jsonnet-fmt script
This the hack/format-jsonnet.sh so it can be called by itself using a
jsonnet-fmt binary in the path, while retaining the option to call it
through make jsonnet-ftm with a temporarily installed binary. This is
preparation to add a jsonnet-ftm CI job.
Also change some formatting changes, jsonnet-fmt generated.

Signed-off-by: Jan Fajerski <jfajersk@redhat.com>