Add recording rule for builds by strategy
Aggregate the number of builds run on a cluster by strategy, and export
the recording rule to Telemeter. This will help the Build API team
determine which build strategies are actively being used on a cluster.

The recording rule will produce a time series with the following
strategy labels:

- docker
- source
- jenkinspipeline
- custom

Updated documentation and telemeter_query