assets: fix instance:node_filesystem_usage:sum
The instance:node_filesystem_usage:sum was previously accounting only
for the `/` mountpoint. This led to some issues were for some nodes, the
query wasn't returning any data. This is caused by the fact that the
`/` mountpoint isn't the rootfs of the host, but a virtual device. The
root of the host is mounted on /host/root in the node-exporter container
and the path to the /host/root is not stripped. The CLI flag to strip
the rootfs path was only added in node-exporter 0.17 and we are shipping
0.16 in OpenShift 3.11.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>