*: use NamespaceSelector instead of explicit allow/deny list
Instead of the `--namespaces` and `deny-namespaces` flags to specify
which namespaces the Prometheus operator should watch/not watch, we
configure the Prometheus resources with the proper namespace selectors.
In practice, the Prometheus `k8s` resource looks for all namespaces with
the `openshift.io/cluster-monitoring=true` label and The Prometheus
`user-workload` resource for all namespaces without the
`openshift.io/cluster-monitoring=true` label.

The goal is to decrease the number of listers/watchers that the operator
needs to maintain which should reduce the load on the API server.
Concretely with the `--namespaces` flag, the operator would create
40 listers/watchers (1 per namespace) per watched object type (secret,
prometheus, statefulset, ...). With namespace selectors, the
operator needs only 1 lister/watcher per object type.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>