Bug 1961081: update pdb apiversion to v1
Pull request #1176 updated the assets managed by the operator
to use the policy/v1 version for all pod disruption budgets.
This was not sufficient to make the operator use the v1 version.

This commit updates the operator code to use the v1 client-go APIs
to make sure no deprecation warnings show in the logs.