pkg/client: Retry on API errors in polling methods
The client package has several methods that poll to wait for certain
conditions to become true, but they all fail immediately when the
initial request(s) to the API server return an error.  This changes
that behavior to ensure the API requests are retried as well, and
simply logs the error returned by the API server at debug level.