operator: more aggressive rate limiting for sync events
This commit changes the rate limiter for events that create a failure in
the sync loop from a `DefaultControllerRateLimiter` to a
`ItemExponentialFailureRateLimiter`. The main difference is that the old
rate limited also has a bucket based limiter. Since we only ever add one
event type (yet) there is no difference. We also go from 5 milliseconds
as the base interval for the exponential backoff to 50 milliseconds with
a maximum of 3 minutes. This should cut down on the number of retries
when other cluster parts are not ready yet, while keeping things
reasonably responsive.

Signed-off-by: Jan Fajerski <jfajersk@redhat.com>