jsonnet: improve recording rules for monitoring
When the Prometheus pods are redeployed and persistent storage is used,
the recording rules used to report telemetry about monitoring would
double count for 5 minutes (at most). It happens because of how
Prometheus handles staleness: the individual metrics would be present
for both the old and new pods if their `instance` labels have changed
(e.g. their IP addresses are different). To avoid this artifact, the
rules aggregate away the `instance` label.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>