jsonnet: Disable cpufreq collector in node_exporter
The node_exporter cpufreq collector seems to be causing high load on
some machines with lots of cores.  This disables it temporarily as a
workaround until the issue can be addressed upstream.  We don't seem
to rely on these metrics anywhere.

See: https://bugzilla.redhat.com/show_bug.cgi?id=1972076