Makefile: update $(JSONNET_VENDOR) dependencies
Upstream sha version for jsonnet code are kept in the
jsonnet/jsonnetfile.lock.json. This file is not a dependency for
generating assets and running make generate can lead to generating assets
from stale jsonnet code.

This commit adds the jsonnet/jsonnetfile.lock.json as a dependency of the
$(JSONNET_VENDOR) Makefile target. With this we ensure the
jsonnet/vendor folder is refetched every time a new version of an
upstream jsonnet dependency has been commited.