Update expression for 'control plane ready' metric to return binary value
The previous version of this expression returned either 1 or null which may have
caused issues with how Prometheus evaluated the expression and hence fired
the related alerts when it should not have.

alert:KubeDeploymentReplicasMismatch: Update alert to account for surges

This change updates the comparison operator in order to account for potential surges.