jsonnet: improve incorrectly spread workload alert
Improve HighlyAvailableWorkloadIncorrectlySpread alert to only fire when
all instances of a workload are on the same node. Ultimately what we
want to catch with this alert are single points of failures and even if
multiple instances end up on the same node, if one of them is correctly
spread, then the workload is highly available.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>