jsonnet: thanosquery: Use HTTP probes as opposed to exec
The exec probe is timing out intermittently. This coud be caused in part
by the timing on the probe not being sufficient, or there may be an underlying
issue with the (exec, cri-o) combo.

This change takes the config for httpGet probes from upstream thanos and proxies
to thanos via the oauth-gateway where we bypass the health check endpoints.

The HTTP probe cannot directly ping thanos since it listens on localhost
and is unreachable from kubelet. This is by design.