Allow users to disable the local Alertmanager
The ACM would like to funnel all alerts from a fleet of OpenShift
clusters through a single external Alertmanager. As a result, they
would like to trim down the monitoring stack and disable the local
alertmanager.

This commit exposes a binary field in the cluster-monitoring-configmap
which allows users to fully disable the built-in alertmanager.