jsonnet: configure mTLS for scrape authentication
This change ensures that all monitoring components (*) configure mTLS
authentication when being scraped by Prometheus. For Thanos Ruler, a
kube-rbc-proxy sidecar container is added because the OAuth proxy
container doesn't support mTLS authn.

(*) the only remaining exception is cluster-monitoring-operator because
it gets deployed by CVO and we have a chicken-and-egg problem to access
the client CA. To solve this, we'd have either to teach kube-rbac-proxy
to load/watch the client CA from the
kube-system/extension-apiserver-authentication configmap or drop
kube-rbac-proxy (e.g. use the HTTP server implementation from
library-go).