*: set Upgradeable status based on pods spreading
The operator will prevent upgrades when stateful pods of the same
component are scheduled on the same node. This work is required to
introduce hard pod anti-affinity and let cluster admins know that they
have to fix the pod spreading before they can proceed since the upgrade
would fail in case of unbalanced spreading.

The operator checks the following resources to set the Upgradeable
condition:
* openshift-monitoring/prometheus-k8s
* openshift-user-workload-monitoring/prometheus-user-workload
* openshift-user-workload-monitoring/thanos-ruler

If pods of the same resource are scheduled on the same node, the
operator would report `Upgradeable: false` with
`reason=WorkloadIncorrectlySpread`.

openshift-monitoring/alertmanager-main isn't concerned yet because we
need first to reduce the number of replicas from 3 to 2 (see
https://bugzilla.redhat.com/show_bug.cgi?id=1955489 for details).

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>
Signed-off-by: Simon Pasquier <spasquie@redhat.com>