*: enable hard anti-affinity + PDB for Prometheus and Ruler
This change introduces back hard pod anti-affinity rules and pod
disruption budgets for Prometheus (both k8s and user-workload) as well
as Thanos Ruler. The cluster monitoring operator updates the
`Upgradeable` condition to false when it detects that the pods aren't
correctly spread to ensure that an upgrade only happens in safe
configurations.

This change fixes a couple of bugs:
* https://bugzilla.redhat.com/show_bug.cgi?id=1933847
* https://bugzilla.redhat.com/show_bug.cgi?id=1949262
* https://bugzilla.redhat.com/show_bug.cgi?id=1955490

It also removes the TestRebalanceWorkloads end-to-end tests because on
a fresh cluster with hard anti-affinity rules, it isn't possible anymore
to trigger a situation where both pods would be scheduled on the same
node.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>