pkg/operator: use alertmanager to set upgradeable
Since alertmanager is meant to have three replicas, the scenario where
it is not correctly spread between multiple nodes is slighly different
then for ther other workloads. In this particular case, even if two
replicas end up on the same node, as long as the third one is on a
different node, then alertmanager should be considered correctly spread
and highly available. Luckily, this is already handled by the current
implementation since we are checking on how many nodes a workload is
spread instead of checking that multiple pods aren't on the same node.
Thus, we can safely re-enable the check for alertmanager pods.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>