*: ensure nodes are uncordoned based on annotation
To make sure that we can track the nodes that are cordoned by cluster we
tag them with the `openshift.io/cluster-monitoring-cordoned` annotation.
If anything happen that would prevent the node from being uncordon,
there is a safety check in the sync loop that verify that none of the
currently unschedulable nodes were cordon by CMO. If any was cordon by
CMO, we make sure to uncordon it.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>