*: delete only PVCs with zonal PVs
Based on the `topology.kubernetes.io/zone` label set by the volume
provisioner, we can know if a PV is zonal or not. Based on that, we
decide whether we should delete the PVC or not since non-zonal PVs can
be access by pods hosted on nodes that are in different zones.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>